Solution for Bracketing Search 4*
This solution is my approach to the given question. I was considering the level of knowledge that i had to use. Though i have to admit that it gave me some troubles to figure it out. So please any comments will be much appreciated about how i can improve the code using only beginner knowledge.